created validation for user delete

+++ index.js
@@ -254,20 +254,36 @@ app.delete('/users/:Username/Movies/:MovieID', passport.authenticate('jwt', { se
 })
 
 // Delete a user by username
-app.delete('/users/:Username', passport.authenticate('jwt', { session: false }), (req, res) => {
-  Users.findOneAndRemove({ Username: req.params.Username })
-    .then((user) => {
-      if (!user) {
-        res.status(400).send(`${req.params.Username} was not found`)
-      } else {
-        res.status(200).send(`${req.params.Username} was deleted.`)
-      }
-    })
-    .catch((err) => {
-      console.error(err)
-      res.status(500).send(`Error: ${err}`)
-    })
-})
+app.delete('/users/:Username', passport.authenticate('jwt', { session: false }),
+  // validation logic here for request
+  // Username: isLength min 5, isAlphanumeric
+  // Password: .not().isEmpty, Email: .isEmail
+  [
+    check('Username', 'Username is required').isLength({ min: 5 }),
+    check('Username', 'Username contains non alphanumeric characters - not allowed.').isAlphanumeric(),
+    check('Password', 'Password is required').not().isEmpty(),
+    check('Email', 'Email does not appear to be valid').isEmail()
+  ], (req, res) => {
+    // check the validation object for errors
+    const errors = validationResult(req)
+
+    if (!errors.isEmpty()) {
+      return res.status(422).json({ errors: errors.array() })
+    }
+
+    Users.findOneAndRemove({ Username: req.params.Username })
+      .then((user) => {
+        if (!user) {
+          res.status(400).send(`${req.params.Username} was not found`)
+        } else {
+          res.status(200).send(`${req.params.Username} was deleted.`)
+        }
+      })
+      .catch((err) => {
+        console.error(err)
+        res.status(500).send(`Error: ${err}`)
+      })
+  })
 
 // listen for requests
 const port = process.env.PORT || 8080
